separate spiritswap products
